moved more responsibility to paths
