+++ autofit/non_linear/abstract_search.py
@@ -1,26 +1,21 @@
 import copy
 import logging
 import multiprocessing as mp
-from os import path
-import os
-import pickle
-import shutil
-from abc import ABC, abstractmethod
-from time import sleep
 import time
-from typing import Dict
+from abc import ABC, abstractmethod
+from os import path
+from time import sleep
 
 import numpy as np
 
 from autoconf import conf
 from autofit import exc
 from autofit.mapper import model_mapper as mm
+from autofit.non_linear import samples as samps
 from autofit.non_linear.initializer import Initializer
 from autofit.non_linear.log import logger
 from autofit.non_linear.paths import Paths, convert_paths
-from autofit.non_linear import samples as samps
 from autofit.non_linear.timer import Timer
-from autofit.text import formatter
 from autofit.text import text_util
 
 
@@ -226,12 +221,6 @@ class NonLinearSearch(ABC):
         and an updated model with free parameters updated to represent beliefs
         produced by this fit.
         """
-
-        try:
-            os.makedirs(self.paths.samples_path)
-        except FileExistsError:
-            pass
-
         self.paths.restore()
         self.setup_log_file()
 
@@ -239,13 +228,15 @@ class NonLinearSearch(ABC):
                 self.paths.has_completed_path
         )) or self.force_pickle_overwrite:
 
-            self.save_model_info(model=model)
-            self.save_parameter_names_file(model=model)
-            self.save_metadata()
-            self.save_info(info=info)
-            self.save_search()
-            self.save_model(model=model)
-            self.move_pickle_files(pickle_files=pickle_files)
+            self.paths.save_model_info(model=model)
+            self.paths.save_parameter_names_file(model=model)
+            self.paths.save_info(info=info)
+            self.paths.save_search(search=self)
+            self.paths.save_model(model=model)
+            self.paths.save_metadata(
+                search_name=type(self).__name__.lower()
+            )
+            self.paths.move_pickle_files(pickle_files=pickle_files)
             analysis.save_attributes_for_aggregator(paths=self.paths)
 
         if not path.exists(self.paths.has_completed_path):
@@ -269,7 +260,7 @@ class NonLinearSearch(ABC):
             samples = self.samples_via_csv_json_from_model(model=model)
 
             if self.force_pickle_overwrite:
-                self.save_samples(samples=samples)
+                self.paths.save_samples(samples=samples)
                 analysis.save_results_for_aggregator(paths=self.paths, samples=samples)
 
         self.paths.zip_remove()
@@ -357,7 +348,7 @@ class NonLinearSearch(ABC):
         samples.write_table(filename=self.paths.samples_file)
         samples.info_to_json(filename=self.paths.info_file)
 
-        self.save_samples(samples=samples)
+        self.paths.save_samples(samples=samples)
 
         try:
             instance = samples.max_log_likelihood_instance
@@ -408,98 +399,6 @@ class NonLinearSearch(ABC):
     @property
     def samples_cls(self):
         raise NotImplementedError()
-
-    def save_model_info(self, model):
-        """Save the model.info file, which summarizes every parameter and prior."""
-        with open(self.paths.file_model_info, "w+") as f:
-            f.write(f"Total Free Parameters = {model.prior_count} \n\n")
-            f.write(model.info)
-
-    def save_parameter_names_file(self, model):
-        """Create the param_names file listing every parameter's label and Latex tag, which is used for *corner.py*
-        visualization.
-
-        The parameter labels are determined using the label.ini and label_format.ini config files."""
-
-        parameter_names = model.model_component_and_parameter_names
-        parameter_labels = model.parameter_labels
-        subscripts = model.subscripts
-        parameter_labels_with_subscript = [f"{label}_{subscript}" for label, subscript in
-                                           zip(parameter_labels, subscripts)]
-
-        parameter_name_and_label = []
-
-        for i in range(model.prior_count):
-            line = formatter.add_whitespace(
-                str0=parameter_names[i], str1=parameter_labels_with_subscript[i], whitespace=70
-            )
-            parameter_name_and_label += [f"{line}\n"]
-
-        formatter.output_list_of_strings_to_file(
-            file=self.paths.file_param_names, list_of_strings=parameter_name_and_label
-        )
-
-    def save_info(self, info):
-        """
-        Save the dataset associated with the phase
-        """
-        with open(path.join(self.paths.pickle_path, "info.pickle"), "wb") as f:
-            pickle.dump(info, f)
-
-    def save_search(self):
-        """
-        Save the seawrch associated with the phase as a pickle
-        """
-        with open(self.paths.make_search_pickle_path(), "w+b") as f:
-            f.write(pickle.dumps(self))
-
-    def save_model(self, model):
-        """
-        Save the model associated with the phase as a pickle
-        """
-        with open(self.paths.make_model_pickle_path(), "w+b") as f:
-            f.write(pickle.dumps(model))
-
-    def save_samples(self, samples):
-        """
-        Save the final-result samples associated with the phase as a pickle
-        """
-
-        with open(self.paths.make_samples_pickle_path(), "w+b") as f:
-            f.write(pickle.dumps(samples))
-
-    def save_metadata(self):
-        """
-        Save metadata associated with the phase, such as the name of the pipeline, the
-        name of the phase and the name of the dataset being fit
-        """
-        with open(path.join(self.paths.make_path(), "metadata"), "a") as f:
-            f.write(self.make_metadata_text())
-
-    def move_pickle_files(self, pickle_files):
-        """
-        Move extra files a user has input the full path + filename of from the location specified to the
-        pickles folder of the Aggregator, so that they can be accessed via the aggregator.
-        """
-        if pickle_files is not None:
-            [shutil.copy(file, self.paths.pickle_path) for file in pickle_files]
-
-    @property
-    def _default_metadata(self) -> Dict[str, str]:
-        """
-        A dictionary of metadata describing this phase, including the pipeline
-        that it's embedded in.
-        """
-        return {
-            "name": self.paths.name,
-            "tag": self.paths.tag,
-            "non_linear_search": type(self).__name__.lower(),
-        }
-
-    def make_metadata_text(self):
-        return "\n".join(
-            f"{key}={value or ''}" for key, value in {**self._default_metadata}.items()
-        )
 
     def remove_state_files(self):
         pass
@@ -552,13 +451,13 @@ class Analysis(ABC):
     def log_likelihood_function(self, instance):
         raise NotImplementedError()
 
-    def visualize(self, paths : Paths, instance, during_analysis):
+    def visualize(self, paths: Paths, instance, during_analysis):
         pass
 
     def save_attributes_for_aggregator(self, paths: Paths):
         pass
 
-    def save_results_for_aggregator(self, paths: Paths, samples : samps.OptimizerSamples):
+    def save_results_for_aggregator(self, paths: Paths, samples: samps.OptimizerSamples):
         pass
